+++ PRTelegramBot/CHANGELOG.md
@@ -4,6 +4,8 @@
 - feature: Добавлены новые события в update типа сообщения. OnPreReplyCommandHandle, OnPostReplyCommandHandle, OnPreDynamicReplyCommandHandle, OnPostDynamicReplyCommandHandle,
     OnPreSlashCommandHandle, OnPostSlashCommandHandle, OnPreInlineCommandHandle, OnPostInlineCommandHandle, OnPreNextStepCommandHandle, OnPostNextStepCommandHandle
 - feature: UpdateExtension добавлены методы IsUserChatId, TryGetChatId
+- feature: Добавлен класс MessageAwaiter, позволяет создавать сообщение заглушку перед обработкой данных и автоматически удалять его после
+- feature: Убраны await для команд reply, slash, inline, dynamicreply чтобы не задерживали обработку других update
 
 -- 2024.06.30 - V0.6.1
 - update: Ядро telegram.bot обновлено с 19 до 21.2.0 версии.